refactor: static 선언 추가
