+++ src/test/java/racingcar/utils/ValidatorTest.java
@@ -8,32 +8,24 @@ import static racingcar.model.constants.Error.NON_COMMA;
 import static racingcar.model.constants.Error.NUMBER_NOT_INTEGER;
 import static racingcar.model.constants.Error.NUMBER_RANGE_INVALID;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
 class ValidatorTest {
-    private Validator validator;
-
-    @BeforeEach
-    void setUp() {
-        validator = new ValidatorImp();
-    }
-
     @ParameterizedTest
     @DisplayName("자동차 이름 입력에 콤마가 없을 때, 길이를 초과할 때, 중복이 있을 때 예외 발생을 통합 테스트 합니다.")
     @ValueSource(strings = {"pobi woni jun", "pobi123,woni,jun", "pobi,pobi,woni,jun"})
     void 자동차_이름_입력_통합_예외(String input) {
-        assertThrows(IllegalArgumentException.class, () -> validator.validateCarName(input));
+        assertThrows(IllegalArgumentException.class, () -> Validator.validateCarName(input));
     }
 
     @ParameterizedTest
     @DisplayName("시도 횟수가 정수가 아닐 때, 시도 횟수가 1보다 작을 때 예외 발생을 통합 테스트 합니다.")
     @ValueSource(strings = {"3.5", "세번", "0", "-1"})
     void 시도_횟수_입력_통합_예외(String input) {
-        assertThrows(IllegalArgumentException.class, () -> validator.validateTrial(input));
+        assertThrows(IllegalArgumentException.class, () -> Validator.validateTrial(input));
     }
 
     @Test
@@ -41,9 +33,9 @@ class ValidatorTest {
     void 자동차_이름_입력_콤마_예외() {
         String input = "pobi woni jun";
 
-        assertThatThrownBy(() -> validator.validateCarName(input))
-              .isInstanceOf(IllegalArgumentException.class)
-              .hasMessageContaining(NON_COMMA.getMessage());
+        assertThatThrownBy(() -> Validator.validateCarName(input))
+                .isInstanceOf(IllegalArgumentException.class)
+                .hasMessageContaining(NON_COMMA.getMessage());
     }
 
     @Test
@@ -51,7 +43,7 @@ class ValidatorTest {
     void 자동차_이름_입력_길이_예외() {
         String input = "pobi123,woni,jun";
 
-        assertThatThrownBy(() -> validator.validateCarName(input))
+        assertThatThrownBy(() -> Validator.validateCarName(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(LENGTH_INVALID.getMessage());
     }
@@ -61,7 +53,7 @@ class ValidatorTest {
     void 자동차_이름_입력_중복_예외() {
         String input = "pobi,pobi,woni,jun";
 
-        assertThatThrownBy(() -> validator.validateCarName(input))
+        assertThatThrownBy(() -> Validator.validateCarName(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(DUPLICATE.getMessage());
     }
@@ -71,7 +63,7 @@ class ValidatorTest {
     void 시도_횟수_입력_정수_예외() {
         String input = "3.5";
 
-        assertThatThrownBy(() -> validator.validateTrial(input))
+        assertThatThrownBy(() -> Validator.validateTrial(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(NUMBER_NOT_INTEGER.getMessage());
     }
@@ -81,7 +73,7 @@ class ValidatorTest {
     void 시도_횟수_입력_범위_예외() {
         String input = "0";
 
-        assertThatThrownBy(() -> validator.validateTrial(input))
+        assertThatThrownBy(() -> Validator.validateTrial(input))
                 .isInstanceOf(IllegalArgumentException.class)
                 .hasMessageContaining(NUMBER_RANGE_INVALID.getMessage());
     }
